Add support for assembling the global matrix and vector using the nonzeros method. This complements the already available element coloring method.
It appears that we get very good performance when we do the assembly (note that the local matrices and vectors are computed separately with this method), but with the cost of large memory consumption that presents us currently to run big simulations.
Particularly the current implementation requires to allocate very large buffers but OpenCL has a limit of 1536 MB per buffer. Allocating more than that seems to prevent the kernel to run.