Make the nonzeros assembly method work with larger simulations by segmenting the kernels enqueuing into several passes. This makes the buffers storing the reduction matrix smaller.
However setUpISMIP_HOM_A010_GPU_NonZeros_dense3 is still too big to run with the nonzeros method since the buffer storing the local matrices is very big and seems to prevent the kernel for the matrix assembly to run. We will need a better solution, but for now
setUpISMIP_HOM_A010_GPU_NonZeros_dense3 only works in single precision with this method.